+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 - Windows paths support
 - Dev services are launched and cleaned up correctly on Windows
+- Dev pages configured with `dank.config.ts` `devPages` option work with
+  both a string webpage mapping and a `DevPageMapping` configuration
 
 ## [v0.0.4] - 2026-02-04